Fix import failure on some PowerShell Core installs by adjusting `System.Management.Automation` references and target frameworks (#23)
* Update module version and PowerShell SDK dependency

Bumped dbatools module version to 2025.8.1 Replaced System.Management.Automation with Microsoft.PowerShell.SDK 7.4.0 in the .csproj to update PowerShell integration.

* Refactor package references by target framework

Hopefully this continues to work on PSv3

Moved System.Threading.Tasks.Dataflow and Microsoft.PowerShell.SDK package references to framework-specific ItemGroups for net472 and net8.0. This allows for correct versioning and dependency management per target framework.

* Bump module version to 2025.8.1 and update copyright

Updated the ModuleVersion to 2025.8.1 and changed the copyright year to 2025 in dbatools.library.psd1.